+++ lib/assets/javascripts/new-dashboard/components/Subscriptions/SubscriptionActions.vue
@@ -13,7 +13,7 @@
       </SubscriptionButtonTooltip>
       <SubscriptionButtonTooltip v-else-if="dataset.sync_status === 'unsynced' && dataset.status === 'active'">
         <button type="button" class="u-mr--8 u-flex u-flex__align--center u-flex__justify--center">
-          <img src="../../assets/icons/catalog/information-circle.svg" :class="smallClass">
+          <img src="../../assets/icons/catalog/information-not-connected.svg" :class="smallClass">
           <div class="tooltip text bgWhite is-small is-txtSoftGrey">
             <h1>Not connected</h1>
             <p>This dataset is not connected. Please, contact support@carto.com.</p>