+++ spec/models/user_shared_examples.rb
@@ -13,11 +13,10 @@ shared_examples_for "user models" do
     include_context 'users helper'
 
     it "should count tweet imports" do
-      FactoryGirl.create(:search_tweet, user: @user1, retrieved_items: 5)
+      FactoryGirl.create(:search_tweet, user: @carto_user1, retrieved_items: 5)
+      FactoryGirl.create(:search_tweet, user: @carto_user2, retrieved_items: 6)
 
-      FactoryGirl.create(:search_tweet, user: @user2, retrieved_items: 6)
-
-      get_twitter_imports_count_by_user_id(@user1.id).should == 5
+      get_twitter_imports_count_by_user_id(@carto_user1.id).should == 5
     end
   end